Ajusta outros perfis e permissoes

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -42,8 +42,13 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getTopico().getAutor(), "ROLE_MODERADOR")){
+            throw new RegraDeNegocioException("Você não pode editar essa resposta!");
+        }
+
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -69,9 +74,13 @@ public class RespostaService {
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR")){
+            throw new RegraDeNegocioException("Você não pode excluir essa resposta!");
+        }
 
         repository.deleteById(id);
