+++ src/main/java/br/com/forum_hub/domain/topico/TopicoService.java
@@ -3,7 +3,9 @@ package br.com.forum_hub.domain.topico;
 import br.com.forum_hub.domain.curso.CursoService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.infra.security.HierarquiaService;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -12,13 +14,14 @@ import org.springframework.stereotype.Service;
 @Service
 public class TopicoService {
 
-    private final TopicoRepository repository;
-    private final CursoService cursoService;
+    @Autowired
+    private TopicoRepository repository;
 
-    public TopicoService(TopicoRepository repository, CursoService cursoService) {
-        this.repository = repository;
-        this.cursoService = cursoService;
-    }
+    @Autowired
+    private CursoService cursoService;
+
+    @Autowired
+    private HierarquiaService hierarquiaService;
 
     @Transactional
     public Topico cadastrar(DadosCadastroTopico dados, Usuario autor) {
@@ -38,15 +41,25 @@ public class TopicoService {
     }
 
     @Transactional
-    public Topico atualizar(DadosAtualizacaoTopico dados) {
+    public Topico atualizar(DadosAtualizacaoTopico dados, Usuario logado) {
         var topico = buscarPeloId(dados.id());
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR")){
+            throw new RegraDeNegocioException("Você não pode editar esse tópico!");
+        }
+
         var curso = cursoService.buscarPeloId(dados.cursoId());
         return topico.atualizarInformacoes(dados, curso);
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var topico = buscarPeloId(id);
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR")){
+            throw new RegraDeNegocioException("Você não pode excluir esse tópico!");
+        }
+
         if (topico.getStatus() == Status.NAO_RESPONDIDO)
             repository.deleteById(id);
         else
